Move the rest of the children code into the new abstraction, TreeElementChildren.
